requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
googletrans==4.0.0rc1
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.1
hyperframe==5.2.0
idna==2.10
more-itertools==9.1.0
multidict==6.0.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.27.0
openai-whisper==20230124
packaging==23.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tokenizers==0.13.2
torch==1.10.1
tqdm==4.65.0
transformers==4.26.1
typing-extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
